refactor dirs 🗄 ..
